Android widget config list adapter cleanup
